Group font and all text into a struct
This puts all the rendered text as well as the UI font into a single
struct. That lets us to dynamically update the text (e.g. the
inventory list).

It's mostly just a PoC for being able to render new text dynamically
without loading it from the disk every time.